+++ app/src/main/AndroidManifest.xml
@@ -20,7 +20,7 @@
         android:networkSecurityConfig="@xml/network_security_config"
         tools:targetApi="31">
       <activity
-        android:name="com.ably.tracking.demo.publisher.ui.MainActivity"
+        android:name="com.ably.tracking.demo.publisher.presentation.MainActivity"
         android:exported="true"
         android:theme="@style/Theme.AATPublisherDemo">
         <intent-filter>
