+++ app/src/main/java/com/ably/tracking/demo/publisher/presentation/screens/login/LoginScreenState.kt
@@ -1,0 +1,8 @@
+package com.ably.tracking.demo.publisher.presentation.screens.login
+
+data class LoginScreenState(
+    val username: String = "",
+    val password: String = "",
+    val showFetchingSecretsFailedDialog: Boolean = false,
+    val showProgress: Boolean = false
+)
