+++ app/src/main/java/com/ably/tracking/demo/publisher/presentation/screens/main/MainScreenState.kt
@@ -1,0 +1,3 @@
+package com.ably.tracking.demo.publisher.presentation.screens.main
+
+data class MainScreenState(val orders: List<OrderViewItem> = emptyList())
