+++ app/src/main/java/com/ably/tracking/demo/publisher/presentation/theme/Color.kt
@@ -1,0 +1,7 @@
+package com.ably.tracking.demo.publisher.presentation.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Primary = Color(0xFF03020D)
+val PrimaryVariant = Color(0xFF000000)
+val Accent = Color(0xFFDC0C0C)
