--- app/src/main/java/com/ably/tracking/demo/publisher/ui/login/LoginScreenState.kt
@@ -1,8 +1,0 @@
-package com.ably.tracking.demo.publisher.ui.login
-
-data class LoginScreenState(
-    val username: String = "",
-    val password: String = "",
-    val showFetchingSecretsFailedDialog: Boolean = false,
-    val showProgress: Boolean = false
-)
