--- app/src/main/java/com/ably/tracking/demo/publisher/ui/main/MainScreenState.kt
@@ -1,3 +1,0 @@
-package com.ably.tracking.demo.publisher.ui.main
-
-data class MainScreenState(val orders: List<OrderViewItem> = emptyList())
