--- app/src/main/java/com/ably/tracking/demo/publisher/ui/main/OrderViewItem.kt
@@ -1,10 +1,0 @@
-package com.ably.tracking.demo.publisher.ui.main
-
-import androidx.annotation.StringRes
-
-data class OrderViewItem(
-    val name: String,
-    @StringRes val state: Int,
-    val onTrackClicked: () -> Unit,
-    val onRemoveClicked: () -> Unit
-)
