--- app/src/main/java/com/ably/tracking/demo/publisher/ui/navigation/Navigator.kt
@@ -1,12 +1,0 @@
-package com.ably.tracking.demo.publisher.ui.navigation
-
-import android.net.Uri
-import androidx.annotation.StringRes
-
-interface Navigator {
-    fun openMain()
-    fun openSettings()
-    fun goBack()
-    fun share(@StringRes logShareHeader: Int, shareFiles: List<Uri>)
-    fun navigateToAppSettingsScreen()
-}
