--- app/src/main/java/com/ably/tracking/demo/publisher/ui/navigation/Routes.kt
@@ -1,16 +1,0 @@
-package com.ably.tracking.demo.publisher.ui.navigation
-
-sealed interface Routes {
-
-    object Login : Routes {
-        const val path = "login/"
-    }
-
-    object Main : Routes {
-        const val path = "main/"
-    }
-
-    object Settings : Routes {
-        const val path = "settings/"
-    }
-}
