--- app/src/main/java/com/ably/tracking/demo/publisher/ui/widget/TextButton.kt
@@ -1,23 +1,0 @@
-package com.ably.tracking.demo.publisher.ui.widget
-
-import androidx.annotation.StringRes
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Button
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun TextButton(@StringRes text: Int, onClick: () -> Unit) {
-    Button(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 8.dp),
-        onClick = onClick
-    ) {
-        Text(stringResource(text))
-    }
-}
